# Vectorize CPU ATen mean kernel for BF16 & FP16 dtypes