--- a/aten/src/ATen/native/ReduceOps.cpp
+++ b/aten/src/ATen/native/ReduceOps.cpp
@@ -1339,12 +1339,12 @@ TORCH_IMPL_FUNC(mean_out)
     }
     auto& result_mut = const_cast<Tensor&>(result);
     if  (dtype == kHalf || dtype == kBFloat16) {
-      // For accuracy reasons, BF16 mean should be computed by following
-      // this approach (FP16 would also use a similar approach):
-      //  cast_fp32->sum ->div-> cast_bf16
+      // For accuracy reasons, BF16/FP16 mean should be computed by following
+      // this approach:
+      //  cast_fp32->sum ->div-> cast_to_bf16_or_fp16
       //
       // This approach results in one extra pass over all the
-      // elements of the output tensor, but the overhead is amortized by
+      // elements of the output tensor, but the overhead may be amortized by
       // vectorization. Such an approach is necessary because
       // cast_fp32->sum-> cast_bf16->cast_fp32->div-> cast_bf16
       // does not produce as precise results.
@@ -1354,14 +1354,14 @@ TORCH_IMPL_FUNC(mean_out)
                                  /*pin_memory=*/false,
                                  /*non_blocking=*/false,
                                  /*copy=*/false,
-                                 c10::nullopt);
+                                 /*memory_format=*/c10::nullopt);
       // a copy of self (input tensor) will be implicitly cast to FP32.
-      // This results in an extra pass over the input array but maybe in the
+      // This results in an extra pass over the input tensor but maybe in the
       // future, temporal locality could be leveraged such that for computing
       // sum, the BF16/FP16 input tensor would be read only once.
       // That would probably require some templatization/special-casing in
       // binary_kernel_reduce_vec(), TensorIteratorBase::for_each(), and
-      // TensorIteratorBase::serial_for_each
+      // TensorIteratorBase::serial_for_each()
       at::sum_out(
           result_mut, self, opt_dim, keepdim, ScalarType::Float).div_(dim_prod);
       // cast result_mut back to BF16 or FP16
@@ -1371,7 +1371,7 @@ TORCH_IMPL_FUNC(mean_out)
                                  /*pin_memory=*/false,
                                  /*non_blocking=*/false,
                                  /*copy=*/false,
-                                 c10::nullopt);
+                                 /*memory_format=*/c10::nullopt);
     } else {
       at::sum_out(result_mut, self, opt_dim, keepdim, dtype).div_(dim_prod);
     }