--- a/aten/src/ATen/native/ReduceOps.cpp
+++ b/aten/src/ATen/native/ReduceOps.cpp
@@ -1338,44 +1338,49 @@ TORCH_IMPL_FUNC(mean_out)
       }
     }
     auto& result_mut = const_cast<Tensor&>(result);
-    if  (dtype == kHalf || dtype == kBFloat16) {
-      // For accuracy reasons, BF16/FP16 mean should be computed by following
-      // this approach:
-      //  cast_fp32->sum ->div-> cast_to_bf16_or_fp16
-      //
-      // This approach results in one extra pass over all the
-      // elements of the output tensor, but the overhead may be amortized by
-      // vectorization. Such an approach is necessary because
-      // cast_fp32->sum-> cast_bf16->cast_fp32->div-> cast_bf16
-      // does not produce as precise results.
-      result_mut = result_mut.to(ScalarType::Float,
-                                 result_mut.layout(),
-                                 result_mut.device(),
-                                 /*pin_memory=*/false,
-                                 /*non_blocking=*/false,
-                                 /*copy=*/false,
-                                 /*memory_format=*/c10::nullopt);
-      // a copy of self (input tensor) will be implicitly cast to FP32.
-      // This results in an extra pass over the input tensor but maybe in the
-      // future, temporal locality could be leveraged such that for computing
-      // sum, the BF16/FP16 input tensor would be read only once.
-      // That would probably require some templatization/special-casing in
-      // binary_kernel_reduce_vec(), TensorIteratorBase::for_each(), and
-      // TensorIteratorBase::serial_for_each()
-      at::sum_out(
-          result_mut, self, opt_dim, keepdim, ScalarType::Float).div_(dim_prod);
-      // cast result_mut back to BF16 or FP16
-      result_mut = result_mut.to(dtype,
-                                 self.layout(),
-                                 self.device(),
-                                 /*pin_memory=*/false,
-                                 /*non_blocking=*/false,
-                                 /*copy=*/false,
-                                 /*memory_format=*/c10::nullopt);
-    } else {
-      at::sum_out(result_mut, self, opt_dim, keepdim, dtype).div_(dim_prod);
+    switch(dtype) {
+      case kHalf:
+      case kBFloat16:
+        // For accuracy reasons, BF16 mean should be computed by following
+        // this approach (FP16 would also use a similar approach):
+        //  cast_fp32->sum ->div-> cast_bf16
+        //
+        // This approach results in one extra pass over all the
+        // elements of the output tensor, but that overhead is amortized by
+        // vectorization. Such an approach is necessary because
+        // cast_fp32->sum-> cast_bf16->cast_fp32->div-> cast_bf16
+        // does not produce as precise results.
+        result_mut = result_mut.to(ScalarType::Float,
+                                   result_mut.layout(),
+                                   result_mut.device(),
+                                   /*pin_memory=*/false,
+                                   /*non_blocking=*/false,
+                                   /*copy=*/false,
+                                   /*memory_format=*/at::MemoryFormat::Preserve);
+        // a copy of self (input tensor) will be implicitly cast to FP32.
+        // This results in an extra pass over the input tensor but maybe in the
+        // future, temporal locality could be leveraged such that for computing
+        // sum, the BF16/FP16 input tensor would be read only once.
+        // That would probably require some templatization/special-casing in
+        // binary_kernel_reduce_vec(), TensorIteratorBase::for_each(), and
+        // TensorIteratorBase::serial_for_each()
+        at::sum_out(
+            result_mut, self, opt_dim, keepdim, ScalarType::Float).div_(dim_prod);
+        // cast result_mut back to BF16 or FP16
+        result_mut = result_mut.to(dtype,
+                                   self.layout(),
+                                   self.device(),
+                                   /*pin_memory=*/false,
+                                   /*non_blocking=*/false,
+                                   /*copy=*/false,
+                                   /*memory_format=*/at::MemoryFormat::Preserve);
+        break;
+      default:
+        // not BF16 & not FP16
+        at::sum_out(result_mut, self, opt_dim, keepdim, dtype).div_(dim_prod);
     }
   } else {
+    // device is not CPU
     auto iter = at::meta::make_reduction_from_out_ty(
         self, result, opt_dim, keepdim, dtype);
     if (iter.numel() == 0) {