--- a/aten/src/ATen/native/ReduceOps.cpp
+++ b/aten/src/ATen/native/ReduceOps.cpp
@@ -1348,13 +1348,7 @@ TORCH_IMPL_FUNC(mean_out)
         // the following code-flow -
         // cast_fp32 -> sum -> cast_bf16 -> cast_fp32 -> div -> cast_bf16,
         // which, in turn, does not produce as accurate results.
-        result_mut = result_mut.to(ScalarType::Float,
-                                   result_mut.layout(),
-                                   result_mut.device(),
-                                   /*pin_memory=*/false,
-                                   /*non_blocking=*/false,
-                                   /*copy=*/false,
-                                   /*memory_format=*/at::MemoryFormat::Preserve);
+        result_mut = result_mut.to(ScalarType::Float);
 
         // self (input tensor) will initially be cast to FP32 in sum_out.
         // This results in having to read that FP32 tensor again, but maybe in
@@ -1362,20 +1356,15 @@ TORCH_IMPL_FUNC(mean_out)
         // that intermediate FP32 tensor. That approach would probably require
         // some modifications in binary_kernel_reduce_vec(),
         // TensorIteratorBase::for_each(), and
-        // TensorIteratorBase::serial_for_each().
+        // TensorIteratorBase::serial_for_each(), apart from sum kernel for CPU.
         at::sum_out(
             result_mut, self, opt_dim, keepdim, ScalarType::Float).div_(dim_prod);
 
         // Cast result_mut back to BF16 or FP16.
-        // A digression - the copy argument to at::native::to would be ignored,
-        // so it might as well have been true in both the at::native::to calls here.
-        result_mut = result_mut.to(dtype,
-                                   self.layout(),
-                                   self.device(),
-                                   /*pin_memory=*/false,
-                                   /*non_blocking=*/false,
-                                   /*copy=*/false,
-                                   /*memory_format=*/at::MemoryFormat::Preserve);
+        // A digression - the copy argument to TensorBase::to would be ignored,
+        // so it might as well have been true in both the TensorBase::to calls
+        // here.
+        result_mut = result_mut.to(dtype);
         break;
       default:
         // not BF16 & not FP16