--- a/aten/src/ATen/native/ReduceOps.cpp
+++ b/aten/src/ATen/native/ReduceOps.cpp
@@ -1347,10 +1347,7 @@ TORCH_IMPL_FUNC(mean_out)
     // which, in turn, does not produce as accurate results.
     bool is_half_type = (dtype == kHalf || dtype == kBFloat16);
     auto sum_out_dtype = is_half_type ? ScalarType::Float : dtype;
-    // TensorBase::to is (sort of a) no-op for all dtypes besides FP16 & BF16.
-    // But we incur a function-call overhead at the expense of simplifying code.
-    // We could add an if condition here to avoid an unnecessary function-call.
-    result_mut = result_mut.to(sum_out_dtype);
+    result_mut = is_half_type ? result_mut.to(sum_out_dtype) : result_mut;
     // If dtype is FP16 or BF16, self (input tensor) will initially be cast to
     // FP32 in sum_out. This results in having to read that FP32 tensor again,
     // but maybe in the future, we could revise the implementation to not
@@ -1360,10 +1357,7 @@ TORCH_IMPL_FUNC(mean_out)
     // TensorIteratorBase::serial_for_each(), apart from sum kernel for CPU.
     at::sum_out(result_mut, self, opt_dim, keepdim, sum_out_dtype).div_(dim_prod);
     // After sum & div, cast result_mut back to BF16 or FP16, if required.
-    // TensorBase::to is (sort of a) no-op for other dtypes, but we incur a
-    // function-call overhead at the expense of simplifying code. We could add
-    // an if condition here to avoid unnecessary function calls.
-    result_mut = result_mut.to(dtype);
+    result_mut = is_half_type ? result_mut.to(dtype) : result_mut;
   } else {
     // device is not CPU
     auto iter = at::meta::make_reduction_from_out_ty(