@sanchitintel 2023-12-09
Hi @malfet, can you please help review this PR?
eager-mode `mean` kernel for BF16/FP16 datatypes has been unvectorized since March, and since it's called by `adaptive_avg_pool2d` in models such as resnet50, their performance with AMP is poor.
This PR fixes the issue by vectorizing `mean` computation for BF16 & FP16 dtypes as well.

Thanks!